refactor: Move `machine` and `utils` dirs to `src` dir
