+++ src/machine/accumulator.js
