+++ src/machine/register.js
