+++ src/machine/stack.js
