+++ src/utils/registers.spec.js
